+++ ChangeLog.rst
@@ -16,6 +16,7 @@ wold-rna-seq-condor now
   * makersemcsv can output either genome or transcriptome summary files
     the filenames have changed slightly to avoid colliding
   * Added support for # as a comment charater in experiment and library files
+  * Now require path to UCSC tools (add ucsc_tools_dir to .htsworkflow.ini)
 
 Release 0.3
 -----------